Used serialization dumps and loads in runner. Moved examples to tests.
