+++ tests/examples.py
